+++ crowdstrike_ztre/__init__.py
@@ -1,1 +1,1 @@
-"""URE CrowdStrike Plugin Package."""
+"""CRE CrowdStrike Plugin Package."""